# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: RailMind Backend
  version: 1.0.0
paths:
...
  /api/v1/assets:
    get:
      tags:
      - infrastructure
      summary: Get Assets
      operationId: get_assets_api_v1_assets_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Page Size
      - name: section_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Section Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_RailwayAsset_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assets/{asset_id}:
    get:
      tags:
      - infrastructure
      summary: Get Asset
      operationId: get_asset_api_v1_assets__asset_id__get
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          title: Asset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_RailwayAsset_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/track-sections:
    get:
      tags:
      - infrastructure
      summary: Get Track Sections
      operationId: get_track_sections_api_v1_track_sections_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Page Size
      - name: corridor_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Corridor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_TrackSection_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/track-sections/{section_id}:
    get:
      tags:
      - infrastructure
      summary: Get Track Section
      operationId: get_track_section_api_v1_track_sections__section_id__get
      parameters:
      - name: section_id
        in: path
        required: true
        schema:
          type: string
          title: Section Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TrackSection_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/corridors:
    get:
      tags:
      - infrastructure
      summary: Get Corridors
      operationId: get_corridors_api_v1_corridors_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_Corridor_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/corridors/{corridor_id}:
    get:
      tags:
      - infrastructure
      summary: Get Corridor
      operationId: get_corridor_api_v1_corridors__corridor_id__get
      parameters:
      - name: corridor_id
        in: path
        required: true
        schema:
          type: string
          title: Corridor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Corridor_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/maintenance/tasks:
    get:
      tags:
      - maintenance
      summary: Get Tasks
      operationId: get_tasks_api_v1_maintenance_tasks_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Page Size
      - name: section_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Section Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_MaintenanceTask_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/maintenance/tasks/{task_id}:
    get:
      tags:
      - maintenance
      summary: Get Task
      operationId: get_task_api_v1_maintenance_tasks__task_id__get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_MaintenanceTask_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/maintenance/defects:
    get:
      tags:
      - maintenance
      summary: Get Defects
      operationId: get_defects_api_v1_maintenance_defects_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_Defect_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/maintenance/defects/{defect_id}:
    get:
      tags:
      - maintenance
      summary: Get Defect
      operationId: get_defect_api_v1_maintenance_defects__defect_id__get
      parameters:
      - name: defect_id
        in: path
        required: true
        schema:
          type: string
          title: Defect Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Defect_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiError:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        httpStatus:
          type: integer
          title: Httpstatus
      type: object
      required:
      - code
      - message
      - httpStatus
      title: ApiError
...
    ApiListResponse_Corridor_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/Corridor'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[Corridor]
...
    ApiListResponse_Defect_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/Defect'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[Defect]
...
    ApiListResponse_MaintenanceTask_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/MaintenanceTask'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[MaintenanceTask]
...
    ApiListResponse_RailwayAsset_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/RailwayAsset'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[RailwayAsset]
...
    ApiListResponse_TrackSection_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/TrackSection'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[TrackSection]
...
    ApiMeta:
      properties:
        timestamp:
          type: string
          title: Timestamp
        requestId:
          type: string
          title: Requestid
        version:
          type: string
          title: Version
      type: object
      required:
      - timestamp
      - requestId
      - version
      title: ApiMeta
...
    ApiResponse_Corridor_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/Corridor'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiResponse[Corridor]
...
    ApiResponse_Defect_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/Defect'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiResponse[Defect]
...
    ApiResponse_MaintenanceTask_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/MaintenanceTask'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiResponse[MaintenanceTask]
...
    ApiResponse_RailwayAsset_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/RailwayAsset'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiResponse[RailwayAsset]
...
    ApiResponse_TrackSection_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/TrackSection'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiResponse[TrackSection]
...
    AssetCategory:
      type: string
      enum:
      - TRACK
      - SIGNAL
      - OHE
      - POINT
      - BRIDGE
      - LEVEL_CROSSING
      title: AssetCategory
...
    Corridor:
      properties:
        corridor_id:
          type: string
          title: Corridor Id
        name:
          type: string
          title: Name
        start_station_id:
          type: string
          title: Start Station Id
        end_station_id:
          type: string
          title: End Station Id
        sections:
          items:
            type: string
          type: array
          title: Sections
      type: object
      required:
      - corridor_id
      - name
      - start_station_id
      - end_station_id
      - sections
      title: Corridor
    Criticality:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - CRITICAL
      title: Criticality
...
    DataSource:
      type: string
      enum:
      - SYSTEM
      - USER
      - MOCK_GENERATOR
      title: DataSource
...
    DataState:
      type: string
      enum:
      - MOCKED
      - LIVE
      title: DataState
...
    Defect:
      properties:
        defect_id:
          type: string
          title: Defect Id
        asset_id:
          type: string
          title: Asset Id
        section_id:
          type: string
          title: Section Id
        defect_type:
          type: string
          title: Defect Type
        description:
          type: string
          title: Description
        severity:
          $ref: '#/components/schemas/DefectSeverity'
        criticality:
          $ref: '#/components/schemas/Criticality'
        detected_at:
          type: string
          format: date-time
          title: Detected At
        detected_by:
          $ref: '#/components/schemas/DefectSource'
        operational_impact:
          type: string
          title: Operational Impact
        is_safety_critical:
          type: boolean
          title: Is Safety Critical
        urgency_hours:
          type: integer
          title: Urgency Hours
        status:
          $ref: '#/components/schemas/DefectStatus'
        linked_task_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Linked Task Id
        department:
          $ref: '#/components/schemas/Department'
        provenance:
          $ref: '#/components/schemas/Provenance'
      type: object
      required:
      - defect_id
      - asset_id
      - section_id
      - defect_type
      - description
      - severity
      - criticality
      - detected_at
      - detected_by
      - operational_impact
      - is_safety_critical
      - urgency_hours
      - status
      - department
      - provenance
      title: Defect
    DefectSeverity:
      type: string
      enum:
      - MINOR
      - MODERATE
      - SEVERE
      - CRITICAL
      title: DefectSeverity
    DefectSource:
      type: string
      enum:
      - INSPECTION
      - AUTOMATED_MONITORING
      - DRIVER_REPORT
      - AEF_SCAN
      - BDMS_IMPORT
      - USER_REPORT
      title: DefectSource
    DefectStatus:
      type: string
      enum:
      - DETECTED
      - ASSESSED
      - LINKED
      - RECTIFIED
      - DEFERRED
      - CLOSED
      title: DefectStatus
...
    Department:
      type: string
      enum:
      - TRACK
      - SIGNALING
      - POWER
      - CIVIL
      - Engineering
      - S&T
      - TRD
      - OHE
      title: Department
...
    DurationMinutes:
      properties:
        expected:
          type: integer
          title: Expected
        minimum:
          type: integer
          title: Minimum
        maximum:
          type: integer
          title: Maximum
      type: object
      required:
      - expected
      - minimum
      - maximum
      title: DurationMinutes
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MaintenanceTask:
      properties:
        task_id:
          type: string
          title: Task Id
        asset_id:
          type: string
          title: Asset Id
        section_id:
          type: string
          title: Section Id
        type:
          $ref: '#/components/schemas/TaskType'
        status:
          $ref: '#/components/schemas/TaskStatus'
        criticality:
          $ref: '#/components/schemas/Criticality'
        department:
          $ref: '#/components/schemas/Department'
        description:
          type: string
          title: Description
        requested_window:
          anyOf:
          - $ref: '#/components/schemas/TimeInterval'
          - type: 'null'
        duration:
          $ref: '#/components/schemas/DurationMinutes'
        requires_power_block:
          type: boolean
          title: Requires Power Block
        requires_traffic_block:
          type: boolean
          title: Requires Traffic Block
        priority_breakdown:
          anyOf:
          - $ref: '#/components/schemas/PriorityBreakdown'
          - type: 'null'
      type: object
      required:
      - task_id
      - asset_id
      - section_id
      - type
      - status
      - criticality
      - department
      - description
      - duration
      - requires_power_block
      - requires_traffic_block
      title: MaintenanceTask
...
    PaginationMeta:
      properties:
        totalItems:
          type: integer
          title: Totalitems
        page:
          type: integer
          title: Page
        pageSize:
          type: integer
          title: Pagesize
        totalPages:
          type: integer
          title: Totalpages
      type: object
      required:
      - totalItems
      - page
      - pageSize
      - totalPages
      title: PaginationMeta
...
    PriorityBreakdown:
      properties:
        safety_score:
          type: integer
          title: Safety Score
        reliability_score:
          type: integer
          title: Reliability Score
        efficiency_score:
          type: integer
          title: Efficiency Score
        total_score:
          type: integer
          title: Total Score
      type: object
      required:
      - safety_score
      - reliability_score
      - efficiency_score
      - total_score
      title: PriorityBreakdown
...
    Provenance:
      properties:
        state:
          $ref: '#/components/schemas/DataState'
        source:
          $ref: '#/components/schemas/DataSource'
        generatedAt:
          type: string
          format: date-time
          title: Generatedat
        generatorVersion:
          anyOf:
          - type: string
          - type: 'null'
          title: Generatorversion
      type: object
      required:
      - state
      - source
      - generatedAt
      title: Provenance
    RailwayAsset:
      properties:
        asset_id:
          type: string
          title: Asset Id
        category:
          $ref: '#/components/schemas/AssetCategory'
        name:
          type: string
          title: Name
        location_section:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Section
        location_station:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Station
        criticality:
          $ref: '#/components/schemas/Criticality'
        is_operational:
          type: boolean
          title: Is Operational
      type: object
      required:
      - asset_id
      - category
      - name
      - criticality
      - is_operational
      title: RailwayAsset
...
    TaskStatus:
      type: string
      enum:
      - PENDING
      - SCHEDULED
      - IN_PROGRESS
      - COMPLETED
      - CANCELLED
      title: TaskStatus
    TaskType:
      type: string
      enum:
      - INSPECTION
      - REPAIR
      - REPLACEMENT
      - PREVENTIVE
      - CORRECTIVE
      - EMERGENCY
      title: TaskType
    TimeInterval:
      properties:
        start:
          type: string
          format: date-time
          title: Start
        end:
          type: string
          format: date-time
          title: End
      type: object
      required:
      - start
      - end
      title: TimeInterval
    TrackSection:
      properties:
        section_id:
          type: string
          title: Section Id
        name:
          type: string
          title: Name
        start_station_id:
          type: string
          title: Start Station Id
        end_station_id:
          type: string
          title: End Station Id
        length_km:
          type: number
          title: Length Km
        max_speed_kmh:
          type: integer
          title: Max Speed Kmh
        is_electrified:
          type: boolean
          title: Is Electrified
        is_bidirectional:
          type: boolean
          title: Is Bidirectional
        track_count:
          type: integer
          title: Track Count
      type: object
      required:
      - section_id
      - name
      - start_station_id
      - end_station_id
      - length_km
      - max_speed_kmh
      - is_electrified
      - is_bidirectional
      - track_count
      title: TrackSection
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...